refactor: 멤버 도메인 패키지 변경에 따른 에러 수정 및 코드 포맷
