+++ src/main/java/com/flint/flint/club/domain/comment/ClubComment.java
@@ -1,10 +1,9 @@
 package com.flint.flint.club.domain.comment;
 
 import com.flint.flint.club.domain.main.Club;
-import com.flint.flint.member.domain.Member;
+import com.flint.flint.member.domain.main.Member;
 import jakarta.persistence.*;
 import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -13,6 +12,7 @@ import java.util.UUID;
 
 /**
  * Club Comment Entity Class
+ *
  * @author 김기현
  * @since 2023-08-03
  */
@@ -20,22 +20,27 @@ import java.util.UUID;
 @NoArgsConstructor
 @Getter
 public class ClubComment {
-    @Id @GeneratedValue(strategy = GenerationType.UUID) @Column(name = "club_comment_id")
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    @Column(name = "club_comment_id")
     private UUID id;
     @Column(name = "club_comment_contents")
     private String contents;
-    @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "club_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "club_id")
     private Club club;
 
     // comment 순환 참조
-    @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "club_comment_parent_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "club_comment_parent_id")
     private ClubComment commentParent;
 
     @OneToMany(mappedBy = "commentParent", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<ClubComment> commentChildren;
 
     // 유저
-    @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "member_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
     private Member member;
 
     @Builder
