+++ src/main/java/com/flint/flint/club/domain/comment/ClubCommentLike.java
@@ -1,9 +1,8 @@
 package com.flint.flint.club.domain.comment;
 
-import com.flint.flint.member.domain.Member;
+import com.flint.flint.member.domain.main.Member;
 import jakarta.persistence.*;
 import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -11,6 +10,7 @@ import java.util.UUID;
 
 /**
  * Club Comment Like Class
+ *
  * @author 김기현
  * @since 2023-08-05
  */
@@ -18,13 +18,17 @@ import java.util.UUID;
 @Getter
 @NoArgsConstructor
 public class ClubCommentLike {
-    @Id @GeneratedValue(strategy = GenerationType.UUID) @Column(name = "club_comment_like_id")
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    @Column(name = "club_comment_like_id")
     private UUID id;
-    @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "club_comment_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "club_comment_id")
     private ClubComment comment;
 
     // 유저
-    @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "member_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
     private Member member;
 
     @Builder
