+++ src/main/java/com/flint/flint/club/domain/main/ClubLike.java
@@ -1,18 +1,14 @@
 package com.flint.flint.club.domain.main;
 
-
-import com.flint.flint.club.domain.main.Club;
-import com.flint.flint.member.domain.Member;
+import com.flint.flint.member.domain.main.Member;
 import jakarta.persistence.*;
 import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.util.UUID;
-
 /**
  * Club Like Class
+ *
  * @author 김기현
  * @since 2023-08-05
  */
@@ -20,13 +16,17 @@ import java.util.UUID;
 @Getter
 @NoArgsConstructor
 public class ClubLike {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY) @Column(name = "club_id")
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "club_id")
     private Long id;
-    @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "memeber_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "memeber_id")
     private Club club;
 
     // 유저
-    @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "member_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
     private Member member;
 
     @Builder
