+++ src/main/java/com/flint/flint/club/domain/main/MemberInClub.java
@@ -1,7 +1,7 @@
 package com.flint.flint.club.domain.main;
 
 import com.flint.flint.club.domain.spec.MemberJoinStatus;
-import com.flint.flint.member.domain.Member;
+import com.flint.flint.member.domain.main.Member;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
@@ -11,13 +11,17 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class MemberInClub {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
-    @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "club_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "club_id")
     private Club club;
-    @ManyToOne(fetch = FetchType.LAZY) @JoinColumn(name = "member_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
     private Member member;
-    @Enumerated(EnumType.STRING) @Column(name = "member_join_status")
+    @Enumerated(EnumType.STRING)
+    @Column(name = "member_join_status")
     private MemberJoinStatus status;
 
     @Builder
